+++ src/main/java/artifacts/item/curio/belt/HeliumFlamingoItem.java
@@ -1,7 +1,6 @@
 package artifacts.item.curio.belt;
 
 import artifacts.Artifacts;
-import artifacts.components.SwimAbilityComponent;
 import artifacts.init.Components;
 import artifacts.init.SoundEvents;
 import artifacts.item.curio.TrinketArtifactItem;
@@ -20,10 +19,17 @@ import net.minecraft.server.network.ServerGamePacketListenerImpl;
 import net.minecraft.world.InteractionResult;
 import net.minecraft.world.entity.player.Player;
 
+import java.util.Collections;
+import java.util.List;
+
 
 public class HeliumFlamingoItem extends TrinketArtifactItem {
 
 	public static final ResourceLocation C2S_AIR_SWIMMING_ID = Artifacts.id("c2s_air_swimming");
+
+	// TODO: config
+	public static final int MAX_FLIGHT_TIME = 150;
+	public static final int RECHARGE_TIME = 300;
 
 	public HeliumFlamingoItem() {
 		super();
@@ -32,18 +38,13 @@ public class HeliumFlamingoItem extends TrinketArtifactItem {
 	}
 
 	private static InteractionResult onPlayerSwim(Player player) {
-		SwimAbilityComponent swimAbilities = Components.SWIM_ABILITIES.get(player);
-
-		if (swimAbilities.isSwimming()) {
-			return InteractionResult.SUCCESS;
-		}
-
-		return InteractionResult.PASS;
+		return Components.SWIM_ABILITIES.get(player).isSwimming() ? InteractionResult.SUCCESS : InteractionResult.PASS;
 	}
 
-	private static void handleAirSwimmingPacket(MinecraftServer server, ServerPlayer player, ServerGamePacketListenerImpl handler, FriendlyByteBuf buf, PacketSender responseSender) {
+	private static void handleAirSwimmingPacket(MinecraftServer server, ServerPlayer player, ServerGamePacketListenerImpl handler, FriendlyByteBuf buf, PacketSender packetSender) {
 		boolean shouldSwim = buf.readBoolean();
-		server.execute(() -> Components.SWIM_ABILITIES.get(player).setSwimming(shouldSwim));
+		server.execute(() -> Components.SWIM_ABILITIES.maybeGet(player)
+				.ifPresent(swimAbilities -> swimAbilities.setSwimming(shouldSwim)));
 	}
 
 	@Override
@@ -57,14 +58,8 @@ public class HeliumFlamingoItem extends TrinketArtifactItem {
 	}
 
 	@Override
-	protected Object[] getTooltipDescriptionArguments() {
+	protected List<String> getTooltipDescriptionArguments() {
 		String translationKey = Minecraft.getInstance().options.keySprint.saveString();
-		return new Object[] {Language.getInstance().getOrDefault(translationKey)};
+		return Collections.singletonList(Language.getInstance().getOrDefault(translationKey));
 	}
-
-/*	public static boolean isFlying(LivingEntity entity) {
-		return TrinketsHelper.isEquipped(Items.HELIUM_FLAMINGO, entity)
-				&& entity instanceof PlayerEntity
-				&& Components.SWIM_ABILITIES.get((PlayerEntity) entity).isAirSwimming();
-	}*/
 }