+++ src/main/java/artifacts/item/curio/belt/UniversalAttractorItem.java
@@ -5,8 +5,8 @@ import artifacts.item.curio.TrinketArtifactItem;
 import artifacts.mixin.mixins.accessors.ItemEntityAccessor;
 import dev.emi.trinkets.api.SlotGroups;
 import dev.emi.trinkets.api.Slots;
+import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.item.ItemEntity;
-import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.phys.AABB;
 import net.minecraft.world.phys.Vec3;
@@ -16,16 +16,17 @@ import java.util.List;
 public class UniversalAttractorItem extends TrinketArtifactItem {
 
 	@Override
-	// magnet logic from Botania, see https://github.com/Vazkii/Botania
-	protected void effectTick(Player player, ItemStack stack) {
-		Vec3 playerPos = player.position().add(0, 0.75, 0);
+	// Magnet logic from Botania, see https://github.com/Vazkii/Botania
+	protected void curioTick(LivingEntity livingEntity, ItemStack stack) {
+		Vec3 playerPos = livingEntity.position().add(0, 0.75, 0);
+		AABB itemRange = new AABB(playerPos, playerPos).inflate(5);
+		List<ItemEntity> items = livingEntity.level.getEntitiesOfClass(ItemEntity.class, itemRange);
 
-		int range = 5;
-		List<ItemEntity> items = player.level.getEntitiesOfClass(ItemEntity.class, new AABB(playerPos.x - range, playerPos.y - range, playerPos.z - range, playerPos.x + range, playerPos.y + range, playerPos.z + range));
 		int pulled = 0;
 		for (ItemEntity item : items) {
-			boolean attractable = Components.DROPPED_ITEM_ENTITY.maybeGet(item).isPresent()
-					&& (!Components.DROPPED_ITEM_ENTITY.get(item).get() || ((ItemEntityAccessor) item).getAge() > 100);
+			boolean attractable = Components.DROPPED_ITEM_ENTITY.maybeGet(item)
+					.map(dropped -> !dropped.get() || ((ItemEntityAccessor) item).getAge() > 100)
+					.orElse(false);
 			if (attractable && item.isAlive() && !item.hasPickUpDelay()) {
 				if (pulled++ > 200) {
 					break;