+++ src/main/java/artifacts/item/curio/feet/AquaDashersItem.java
@@ -1,9 +1,10 @@
 package artifacts.item.curio.feet;
 
-import artifacts.components.SwimAbilityComponent;
 import artifacts.init.Components;
+import artifacts.init.Items;
 import artifacts.item.curio.TrinketArtifactItem;
 import artifacts.trinkets.Slots;
+import artifacts.trinkets.TrinketsHelper;
 import be.florens.expandability.api.fabric.LivingFluidCollisionCallback;
 import dev.emi.trinkets.api.SlotGroups;
 import net.minecraft.tags.FluidTags;
@@ -20,20 +21,25 @@ public class AquaDashersItem extends TrinketArtifactItem {
 	}
 
 	private static boolean onFluidCollision(LivingEntity entity, FluidState fluidState) {
-		if (entity.isSprinting() && entity.fallDistance < 6 && !entity.isUsingItem() && !entity.isCrouching()) {
-			SwimAbilityComponent swimAbilities = Components.SWIM_ABILITIES.get(entity);
-
-			if (!swimAbilities.isWet() && !swimAbilities.isSwimming()) {
-				if (fluidState.is((FluidTags.LAVA))) {
-					if (!entity.fireImmune() && !EnchantmentHelper.hasFrostWalker(entity)) {
-						entity.hurt(DamageSource.HOT_FLOOR, 1);
-					}
-				}
-				return true;
+		if (TrinketsHelper.isEquipped(Items.AQUA_DASHERS, entity) && canSprintOnWater(entity)) {
+			if (fluidState.is((FluidTags.LAVA)) && !entity.fireImmune() && !EnchantmentHelper.hasFrostWalker(entity)) {
+				entity.hurt(DamageSource.HOT_FLOOR, 1);
 			}
+			return true;
 		}
 
 		return false;
+	}
+
+	private static boolean canSprintOnWater(LivingEntity entity) {
+		return Components.SWIM_ABILITIES.maybeGet(entity)
+				.map(swimAbilities -> entity.isSprinting()
+						&& entity.fallDistance < 6
+						&& !entity.isUsingItem()
+						&& !entity.isCrouching()
+						&& !swimAbilities.isWet()
+						&& !swimAbilities.isSwimming())
+				.orElse(false);
 	}
 
 	@Override