+++ src/main/java/artifacts/mixin/mixins/dev/AbstractMinecartContainerMixin.java
@@ -1,0 +1,17 @@
+package artifacts.mixin.mixins.dev;
+
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.entity.vehicle.AbstractMinecartContainer;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Redirect;
+
+@Mixin(AbstractMinecartContainer.class)
+public class AbstractMinecartContainerMixin {
+
+    @Redirect(method = "createMenu(ILnet/minecraft/world/entity/player/Inventory;Lnet/minecraft/world/entity/player/Player;)Lnet/minecraft/world/inventory/AbstractContainerMenu;",
+            at = @At(value = "INVOKE", target = "Lnet/minecraft/world/entity/player/Player;isSpectator()Z"))
+    private boolean spectatorUnlockable(Player player) {
+        return false;
+    }
+}